ahci_xgene: add errata workaround for ATA_CMD_SMART
commit 2a0bdff6b958d1b2:

 ahci_xgene: fix the dma state machine lockup for the IDENTIFY DEVICE PIO mode command.

added a workaround for X-Gene AHCI controller errata. This was done
for all ATA_CMD_ID_ATA commands. The errata also appears to affect
ATA_CMD_SMART commands as well. This was discovered when running
smartd or just smartctl -x. This patch adds a dma engine restart for
ATA_CMD_SMART commands which clears up the issues seen with smartd.

Signed-off-by: Mark Salter <msalter@redhat.com>