arm64: add parking protocol support
This is a first-cut effort at parking protocol support. It is
very much a work in progress (as is the spec it is based on).
This code deviates from the current spec in a number of ways
to work around current firmware issues and issues with kernels
using 64K page sizes.

caveat utilitor

Signed-off-by: Mark Salter <msalter@redhat.com>