arm64: add smp parking protocol support for X-Gene platform
This adds the smp parking protocol support for APM X-Gene platform. Currently
this patch only support putting the cpu core back to parking status. It does
not perform any power management operation on the cpu. It is send to a wait
state where it waits for a wake signal. Upon a which it will rejoin the
collective.